add constants, UserDetailCard, PostItem
