+++ app/src/apps/users/components/UserDetails.tsx
@@ -1,18 +1,19 @@
-import React from 'react';
-import { useQuery } from '@tanstack/react-query';
-import { useParams } from 'react-router-dom';
-import { Card, CardContent, Typography } from '@mui/material';
-import { userApi } from '../services/userApi';
-import { Error } from '@/shared/components/Error';
-import { Loading } from '@/shared/components/Loading';
-import { PostList } from '@/apps/posts/components/PostList';
+import React from "react";
+import { useQuery } from "@tanstack/react-query";
+import { useParams } from "react-router-dom";
+import { Typography } from "@mui/material";
+import { userApi } from "../services/userApi";
+import { Error } from "@/shared/components/Error";
+import { Loading } from "@/shared/components/Loading";
+import { PostList } from "@/apps/posts/components/PostList";
+import { UserDetailCard } from "@/apps/users/components/UserDetailCard";
 
 export const UserDetails: React.FC = () => {
   const { id } = useParams<{ id: string }>();
   const userId = parseInt(id!, 10);
 
   const { data: user, isLoading, isError, error } = useQuery({
-    queryKey: ['user', userId],
+    queryKey: ["user", userId],
     queryFn: () => userApi.getUserById(userId),
   });
 
@@ -22,16 +23,10 @@ export const UserDetails: React.FC = () => {
 
   return (
     <>
-      <Card sx={{ mb: 4 }}>
-        <CardContent>
-          <Typography variant="h5">{user.name}</Typography>
-          <Typography color="textSecondary">{user.email}</Typography>
-          <Typography>Gender: {user.gender}</Typography>
-          <Typography>Status: {user.status}</Typography>
-        </CardContent>
-      </Card>
-      
-      <Typography variant="h6" sx={{ mb: 2 }}>User Posts</Typography>
+      <UserDetailCard user={user} />
+      <Typography variant="h6" sx={{ mb: 2 }}>
+        User Posts
+      </Typography>
       <PostList userId={userId} />
     </>
   );
