+++ app/src/apps/users/constants/assets.ts
@@ -1,0 +1,5 @@
+export const ASSETS = {
+    KING_CROWN_URL: "/src/assets/king-crown.png",
+    QUEEN_CROWN_URL: "/src/assets/queen-crown.png",
+    GUARD_ICON_URL: "/src/assets/guard.png"
+  }
