+++ app/src/shared/components/LoadMore.tsx
@@ -1,0 +1,8 @@
+import { forwardRef } from 'react';
+import styles from './loadMore.module.css';
+
+export const LoadMore = forwardRef<HTMLDivElement>((_, ref) => {
+  return <div ref={ref} className={styles.loadMore} />;
+});
+
+LoadMore.displayName = 'LoadMore';
