+++ app/src/shared/components/loadMore.module.css
@@ -1,0 +1,4 @@
+.loadMore {
+    height: 20px;
+    width: 100%;
+  }
