+++ app/src/shared/constants/pagination.ts
@@ -1,0 +1,4 @@
+export const PAGINATION = {
+    DEFAULT_PAGE: 1,
+    PER_PAGE: 10
+  }
